Move terminal telemetry into a terminalContrib
This is easily isolated in a contrib, simplifying the core terminal further
