+++ src/vs/workbench/contrib/terminal/terminal.all.ts
@@ -34,5 +34,6 @@ import '../terminalContrib/resizeDimensionsOverlay/browser/terminal.resizeDimens
 import '../terminalContrib/sendSequence/browser/terminal.sendSequence.contribution.js';
 import '../terminalContrib/sendSignal/browser/terminal.sendSignal.contribution.js';
 import '../terminalContrib/suggest/browser/terminal.suggest.contribution.js';
+import '../terminalContrib/telemetry/browser/terminal.telemetry.contribution.js';
 import '../terminalContrib/wslRecommendation/browser/terminal.wslRecommendation.contribution.js';
 import '../terminalContrib/voice/browser/terminal.voice.contribution.js';
